Minor code cleanup in preparation for using Model
What isn't shown is I spent 3 hours trying to get the model to work but constantly ran into memory access issues :(

+++ LearnOpenGL/LearnOpenGL/src/Shader.hpp
@@ -9,7 +9,7 @@ using std::string;
 
 namespace Engine
 {
-	enum class ShaderType : unsigned short
+	enum class ShaderType : uint8_t
 	{
 		PROGRAM,
 		VERTEX,
